requirements.txt
google-cloud-bigquery
google-cloud-storage
google-cloud-pubsub
google-cloud-secret-manager
google-auth
typer
cryptography
